Remove unnecessary code on PianoEditor.scrollToNote
There is a bug where if you create the PianoEditor and call edit() right
away then scrollIntoView won't work.

I've tried all sorts window.requestAnimationFrame() shenanigans and
those worked on Edge but not on Firefox...

Better just leave it be for now.
In the final game i will probably just create the editor on startup
and it will be a while after it's edit() method is called...

Although when i release the PianoEditor as a open-source module
i might look into this issue.

Also any kind of async stuff in scrollToNote() might make it more
difficult to implement the future doUndo functionality which will
probably include reverting scrollTop/Left position.

For now to continue using the current "edit sequence on init" approach
for ease of development while creating the Piano Editor I've added a
150ms delay between creating the editor.